Moving hypothesis tests to their own directory and only running on >python 2.6
